extracted expense input content + usage in add expense screen
