+++ composeApp/src/commonMain/kotlin/org/example/project/feature/home/presentation/HomeViewModel.kt
@@ -29,7 +29,7 @@ class HomeViewModel(
     override fun onIntent(intent: HomeIntent) {
         when (intent) {
             HomeIntent.Refresh -> { /* Orchestrator refresh TBD; avoid duplicate initialize */ }
-            HomeIntent.NavigateToAddExpense -> navigationManager.navigateTo(Screen.AddExpense())
+            HomeIntent.NavigateToAddExpense -> navigationManager.navigateTo(Screen.AddExpense)
             is HomeIntent.NavigateToExpenseDetail -> navigationManager.navigateTo(Screen.ExpenseDetail(intent.expenseId))
         }
     }
