+++ composeApp/src/commonMain/kotlin/org/example/project/navigation/Navigator.kt
@@ -157,7 +157,7 @@ private fun ScreenPlaceholder(
         }
 
         if (showNavigationButtons) {
-            Button(onClick = { navigationManager.navigateTo(Screen.AddExpense()) }) {
+            Button(onClick = { navigationManager.navigateTo(Screen.AddExpense) }) {
                 Text("Add Expense")
             }
